// backend/generated/prisma/models/Utilisateur.ts
export type UtilisateurModel = runtime.Types.Result.DefaultSelection<Prisma.$UtilisateurPayload>
// ...
export type $UtilisateurPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Utilisateur"
  objects: {
    comptes: Prisma.$ComptePayload<ExtArgs>[]
    categories: Prisma.$CategoriePayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: number
    nom: string
    prenom: string
    email: string
    motDePasse: string
    refreshToken: string | null
  }, ExtArgs["result"]["utilisateur"]>
  composites: {}
}